created the unblock user service, controller, and route
