+++ src/controllers/blockController.ts
@@ -10,13 +10,25 @@ const blockUser = asyncWrapper(
 
     await blockServices.blockUserService(userId, userToBlockId);
 
-    return res
-      .status(200)
-      .json({
-        status: httpStatusText.SUCCESS,
-        data: { message: "User blocked successfully" },
-      });
+    return res.status(200).json({
+      status: httpStatusText.SUCCESS,
+      data: { message: "User blocked successfully" },
+    });
   }
 );
 
-export { blockUser };
+const deleteBlock = asyncWrapper(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { userId } = req.currentUser!;
+    const { blockId } = req.params;
+
+    await blockServices.deleteBlockService(userId, blockId);
+
+    return res.status(200).json({
+      status: httpStatusText.SUCCESS,
+      data: { message: "User unblocked successfully" },
+    });
+  }
+);
+
+export { blockUser, deleteBlock };
